refactor(auth): extract _resolve_auth from client.py to _auth.py
KaguraClient was the sole consumer of _resolve_auth() and its
_StaticAuth / _OAuthAuth dataclasses. Issue #110 (B) needs the same
resolver from FilesClient, and a third consumer (ResourceClient) is
already on the roadmap. Keeping the resolver in client.py would force
FilesClient → KaguraClient import direction — unnatural since they
sit at the same layer.

Move _resolve_auth, _StaticAuth, _OAuthAuth, and _DEFAULT_MCP_URL
verbatim to a new src/kagura_memory/_auth.py module. client.py imports
the two symbols it actually references (_resolve_auth, _StaticAuth).
The module imports flow one-way: _auth → exceptions / auth.credentials /
config, with no reverse dependency.

No behavior change — the implementation is identical line for line.
Updated two tests' monkeypatch paths from kagura_memory.client.load_config
to kagura_memory._auth.load_config.

Refs #110.
